% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ab01nd.R
\name{ab01nd}
\alias{ab01nd}
\title{ab01nd}
\usage{
ab01nd(jobz, n, m, a, b, tol, ldwork)
}
\description{
Controllable realization for multi-input systems using orthogonal state and input transformations
}
\examples{
  To find a controllable realization for the linear time-invariant
  multi-input system

          dX/dt = A * X + B * U,

  where A and B are N-by-N and N-by-M matrices, respectively,
  which are reduced by this routine to orthogonal canonical form
  using (and optionally accumulating) orthogonal similarity
  transformations.  Specifically, the pair (A, B) is reduced to
  the pair (Ac, Bc),  Ac = Z' * A * Z,  Bc = Z' * B,  given by

          [ Acont     *    ]         [ Bcont ]
     Ac = [                ],   Bc = [       ],
          [   0    Auncont ]         [   0   ]

     and

             [ A11 A12  . . .  A1,p-1 A1p ]         [ B1 ]
             [ A21 A22  . . .  A2,p-1 A2p ]         [ 0  ]
             [  0  A32  . . .  A3,p-1 A3p ]         [ 0  ]
     Acont = [  .   .   . . .    .     .  ],   Bc = [ .  ],
             [  .   .     . .    .     .  ]         [ .  ]
             [  .   .       .    .     .  ]         [ .  ]
             [  0   0   . . .  Ap,p-1 App ]         [ 0  ]

  where the blocks  B1, A21, ..., Ap,p-1  have full row ranks and
  p is the controllability index of the pair.  The size of the
  block  Auncont is equal to the dimension of the uncontrollable
  subspace of the pair (A, B).

#'
}
\references{
\url{http://slicot.org/objects/software/shared/doc/AB01ND.html}
}
