% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ab05md.R
\name{ab05md}
\alias{ab05md}
\title{ab05md}
\usage{
ab05md(uplo, n1, m1, p1, n2, p2, a1, b1, c1, d1, a2, b2, c2, d2, ldwork)
}
\description{
Cascade inter-connection of two systems in state-space form
}
\examples{
  To obtain the state-space model (A,B,C,D) for the cascaded
  inter-connection of two systems, each given in state-space form.

#'
}
\references{
\url{http://slicot.org/objects/software/shared/doc/AB05MD.html}
}
