% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ab05nd.R
\name{ab05nd}
\alias{ab05nd}
\title{ab05nd}
\usage{
ab05nd(n1, m1, p1, n2, alpha, a1, b1, c1, d1, a2, b2, c2, d2, ldwork)
}
\description{
Feedback inter-connection of two systems in state-space form
}
\examples{
  To obtain the state-space model (A,B,C,D) for the feedback
  inter-connection of two systems, each given in state-space form.

#'
}
\references{
\url{http://slicot.org/objects/software/shared/doc/AB05ND.html}
}
