% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ab07nd.R
\name{ab07nd}
\alias{ab07nd}
\title{ab07nd}
\usage{
ab07nd(n, m, a, b, c, d, ldwork)
}
\description{
Inverse of a given linear system
}
\examples{
  To compute the inverse (Ai,Bi,Ci,Di) of a given system (A,B,C,D).

#'
}
\references{
\url{http://slicot.org/objects/software/shared/doc/AB07ND.html}
}
