% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ab08nd.R
\name{ab08nd}
\alias{ab08nd}
\title{ab08nd}
\usage{
ab08nd(equil, n, m, p, a, b, c, d, tol, ldwork)
}
\description{
Construction of a regular pencil for a given system such that its generalized eigenvalues are invariant zeros of the system
}
\examples{
  To construct for a linear multivariable system described by a
  state-space model (A,B,C,D) a regular pencil (A - lambda*B ) which
                                                 f          f
  has the invariant zeros of the system as generalized eigenvalues.
  The routine also computes the orders of the infinite zeros and the
  right and left Kronecker indices of the system (A,B,C,D).

#'
}
\references{
\url{http://slicot.org/objects/software/shared/doc/AB08ND.html}
}
