% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ab09ad.R
\name{ab09ad}
\alias{ab09ad}
\title{ab09ad}
\usage{
ab09ad(dico, job, equil, ordsel, n, m, p, nr, a, b, c, tol, ldwork)
}
\description{
Balance & Truncate model reduction for stable systems
}
\examples{
  To compute a reduced order model (Ar,Br,Cr) for a stable original
  state-space representation (A,B,C) by using either the square-root
  or the balancing-free square-root Balance & Truncate (B & T)
  model reduction method.

#'
}
\references{
\url{http://slicot.org/objects/software/shared/doc/AB09AD.html}
}
