% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ab09ax.R
\name{ab09ax}
\alias{ab09ax}
\title{ab09ax}
\usage{
ab09ax(dico, job, ordsel, n, m, p, nr, a, b, c, tol, ldwork)
}
\description{
Balance & Truncate model reduction for stable systems with state matrix in real Schur canonical form
}
\examples{
  To compute a reduced order model (Ar,Br,Cr) for a stable original
  state-space representation (A,B,C) by using either the square-root
  or the balancing-free square-root Balance & Truncate model
  reduction method. The state dynamics matrix A of the original
  system is an upper quasi-triangular matrix in real Schur canonical
  form. The matrices of the reduced order system are computed using
  the truncation formulas:

       Ar = TI * A * T ,  Br = TI * B ,  Cr = C * T .

#'
}
\references{
\url{http://slicot.org/objects/software/shared/doc/AB09AX.html}
}
