% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ab09bd.R
\name{ab09bd}
\alias{ab09bd}
\title{ab09bd}
\usage{
ab09bd(dico, job, equil, ordsel, n, m, p, nr, a, b, c, d, tol1, tol2, ldwork)
}
\description{
Singular Perturbation Approximation based model reduction for stable systems
}
\examples{
  To compute a reduced order model (Ar,Br,Cr,Dr) for a stable
  original state-space representation (A,B,C,D) by using either the
  square-root or the balancing-free square-root Singular
  Perturbation Approximation (SPA) model reduction method.

#'
}
\references{
\url{http://slicot.org/objects/software/shared/doc/AB09BD.html}
}
