% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ab09md.R
\name{ab09md}
\alias{ab09md}
\title{ab09md}
\usage{
ab09md(dico, job, equil, ordsel, n, m, p, nr, alpha, a, b, c, tol, ldwork)
}
\description{
Balance & Truncate model reduction for the stable part of a system
}
\examples{
  To compute a reduced order model (Ar,Br,Cr) for an original
  state-space representation (A,B,C) by using either the square-root
  or the balancing-free square-root Balance & Truncate (B & T)
  model reduction method for the ALPHA-stable part of the system.

#'
}
\references{
\url{http://slicot.org/objects/software/shared/doc/AB09MD.html}
}
