% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ab09nd.R
\name{ab09nd}
\alias{ab09nd}
\title{ab09nd}
\usage{
ab09nd(
  dico,
  job,
  equil,
  ordsel,
  n,
  m,
  p,
  nr,
  alpha,
  a,
  b,
  c,
  d,
  tol1,
  tol2,
  ldwork
)
}
\description{
Singular perturbation approximation based model reduction for the stable part of a system
}
\examples{
  To compute a reduced order model (Ar,Br,Cr,Dr) for an original
  state-space representation (A,B,C,D) by using either the
  square-root or the balancing-free square-root Singular
  Perturbation Approximation (SPA) model reduction method for the
  ALPHA-stable part of the system.

#'
}
\references{
\url{http://slicot.org/objects/software/shared/doc/AB09ND.html}
}
