% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ab13bd.R
\name{ab13bd}
\alias{ab13bd}
\title{ab13bd}
\usage{
ab13bd(dico, jobn, n, m, p, a, b, c, d, tol)
}
\description{
H2 or L2 norm of a system
}
\examples{
  To compute the H2 or L2 norm of the transfer-function matrix G
  of the system (A,B,C,D). G must not have poles on the imaginary
  axis, for a continuous-time system, or on the unit circle, for
  a discrete-time system. If the H2-norm is computed, the system
  must be stable.

#'
}
\references{
\url{http://slicot.org/objects/software/shared/doc/AB13BD.html}
}
