% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ab13dd.R
\name{ab13dd}
\alias{ab13dd}
\title{ab13dd}
\usage{
ab13dd(dico, jobe, equil, jobd, n, m, p, fpeak, a, e, b, c, d, tol)
}
\description{
L-infinity norm of a state space system
}
\examples{
  To compute the L-infinity norm of a continuous-time or
  discrete-time system, either standard or in the descriptor form,

                                  -1
     G(lambda) = C*( lambda*E - A ) *B + D .

  The norm is finite if and only if the matrix pair (A,E) has no
  eigenvalue on the boundary of the stability domain, i.e., the
  imaginary axis, or the unit circle, respectively. It is assumed
  that the matrix E is nonsingular.

#'
}
\references{
\url{http://slicot.org/objects/software/shared/doc/AB13DD.html}
}
