% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ab13ed.R
\name{ab13ed}
\alias{ab13ed}
\title{ab13ed}
\usage{
ab13ed(n, a, tol)
}
\description{
Estimating the distance from a real matrix to the nearest complex matrix with an eigenvalue on the imaginary axis, using bisection
}
\examples{
  To estimate beta(A), the 2-norm distance from a real matrix A to
  the nearest complex matrix with an eigenvalue on the imaginary
  axis. The estimate is given as

         LOW <= beta(A) <= HIGH,

  where either

         (1 + TOL) * LOW >= HIGH,

  or

         LOW = 0   and   HIGH = delta,

  and delta is a small number approximately equal to the square root
  of machine precision times the Frobenius norm (Euclidean norm)
  of A. If A is stable in the sense that all eigenvalues of A lie
  in the open left half complex plane, then beta(A) is the distance
  to the nearest unstable complex matrix, i.e., the complex
  stability radius.

#'
}
\references{
\url{http://slicot.org/objects/software/shared/doc/AB13ED.html}
}
