% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag08bd.R
\name{ag08bd}
\alias{ag08bd}
\title{ag08bd}
\usage{
ag08bd(equil, l, n, m, p, a, e, b, c, d, tol, ldwork)
}
\description{
Zeros and Kronecker structure of a descriptor system pencil
}
\examples{
  To extract from the system pencil

                    ( A-lambda*E B )
        S(lambda) = (              )
                    (      C     D )

  a regular pencil Af-lambda*Ef which has the finite Smith zeros of
  S(lambda) as generalized eigenvalues. The routine also computes
  the orders of the infinite Smith zeros and determines the singular
  and infinite Kronecker structure of system pencil, i.e., the right
  and left Kronecker indices, and the multiplicities of infinite
  eigenvalues.

#'
}
\references{
\url{http://slicot.org/objects/software/shared/doc/AG08BD.html}
}
