% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mb03rd.R
\name{mb03rd}
\alias{mb03rd}
\title{mb03rd}
\usage{
mb03rd(jobx, sort, n, pmax, a, x, tol)
}
\description{
Reduction of a real Schur form matrix to a block-diagonal form
}
\examples{
  To reduce a matrix A in real Schur form to a block-diagonal form
  using well-conditioned non-orthogonal similarity transformations.
  The condition numbers of the transformations used for reduction
  are roughly bounded by PMAX*PMAX, where PMAX is a given value.
  The transformations are optionally postmultiplied in a given
  matrix X. The real Schur form is optionally ordered, so that
  clustered eigenvalues are grouped in the same block.

#'
}
\references{
\url{http://slicot.org/objects/software/shared/doc/MB03RD.html}
}
