% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mb03vd.R
\name{mb03vd}
\alias{mb03vd}
\title{mb03vd}
\usage{
mb03vd(n, ilo, ihi, a)
}
\description{
Periodic Hessenberg form of a product of p matrices using orthogonal similarity transformations
}
\examples{
  To reduce a product of p real general matrices A = A_1*A_2*...*A_p
  to upper Hessenberg form, H = H_1*H_2*...*H_p, where H_1 is
  upper Hessenberg, and H_2, ..., H_p are upper triangular, by using
  orthogonal similarity transformations on A,

          Q_1' * A_1 * Q_2 = H_1,
          Q_2' * A_2 * Q_3 = H_2,
                 ...
          Q_p' * A_p * Q_1 = H_p.

#'
}
\references{
\url{http://slicot.org/objects/software/shared/doc/MB03VD.html}
}
