% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mb03vy.R
\name{mb03vy}
\alias{mb03vy}
\title{mb03vy}
\usage{
mb03vy(n, ilo, ihi, a, tau, ldwork)
}
\description{
Orthogonal matrices for reduction to periodic Hessenberg form of a product of matrices
}
\examples{
  To generate the real orthogonal matrices Q_1, Q_2, ..., Q_p,
  which are defined as the product of ihi-ilo elementary reflectors
  of order n, as returned by SLICOT Library routine MB03VD:

     Q_j = H_j(ilo) H_j(ilo+1) . . . H_j(ihi-1).

#'
}
\references{
\url{http://slicot.org/objects/software/shared/doc/MB03VY.html}
}
