% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mb05nd.R
\name{mb05nd}
\alias{mb05nd}
\title{mb05nd}
\usage{
mb05nd(n, delta, a, tol)
}
\description{
Matrix exponential and integral for a real matrix
}
\examples{
  To compute

  (a)    F(delta) =  exp(A*delta) and

  (b)    H(delta) =  Int[F(s) ds] from s = 0 to s = delta,

  where A is a real N-by-N matrix and delta is a scalar value.

#'
}
\references{
\url{http://slicot.org/objects/software/shared/doc/MB05ND.html}
}
