% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc01td.R
\name{mc01td}
\alias{mc01td}
\title{mc01td}
\usage{
mc01td(dico, dp, p)
}
\description{
Checking stability of a given real polynomial
}
\examples{
  To determine whether or not a given polynomial P(x) with real
  coefficients is stable, either in the continuous-time or discrete-
  time case.

  A polynomial is said to be stable in the continuous-time case
  if all its zeros lie in the left half-plane, and stable in the
  discrete-time case if all its zeros lie inside the unit circle.

#'
}
\references{
\url{http://slicot.org/objects/software/shared/doc/MC01TD.html}
}
