% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sb01bd.R
\name{sb01bd}
\alias{sb01bd}
\title{sb01bd}
\usage{
sb01bd(dico, n, m, np, alpha, a, b, wr, wi, tol, ldwork)
}
\description{
Pole assignment for a given matrix pair (A,B)
}
\examples{
  To determine the state feedback matrix F for a given system (A,B)
  such that the closed-loop state matrix A+B*F has specified
  eigenvalues.

#'
}
\references{
\url{http://slicot.org/objects/software/shared/doc/SB01BD.html}
}
