% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sb02md.R
\name{sb02md}
\alias{sb02md}
\title{sb02md}
\usage{
sb02md(dico, hinv, uplo, scal, sort, n, a, g, q, ldwork)
}
\description{
Solution of continuous- or discrete-time algebraic Riccati equations (Schur vectors method)
}
\examples{
  To solve for X either the continuous-time algebraic Riccati
  equation
                           -1
     Q + A'*X + X*A - X*B*R  B'*X = 0                            (1)

  or the discrete-time algebraic Riccati equation
                                     -1
     X = A'*X*A - A'*X*B*(R + B'*X*B)  B'*X*A + Q                (2)

  where A, B, Q and R are N-by-N, N-by-M, N-by-N and M-by-M matrices
  respectively, with Q symmetric and R symmetric nonsingular; X is
  an N-by-N symmetric matrix.
                    -1
  The matrix G = B*R  B' must be provided on input, instead of B and
  R, that is, for instance, the continuous-time equation

     Q + A'*X + X*A - X*G*X = 0                                  (3)

  is solved, where G is an N-by-N symmetric matrix. SLICOT Library
  routine SB02MT should be used to compute G, given B and R. SB02MT
  also enables to solve Riccati equations corresponding to optimal
  problems with coupling terms.

  The routine also returns the computed values of the closed-loop
  spectrum of the optimal system, i.e., the stable eigenvalues
  lambda(1),...,lambda(N) of the corresponding Hamiltonian or
  symplectic matrix associated to the optimal problem.

#'
}
\references{
\url{http://slicot.org/objects/software/shared/doc/SB02MD.html}
}
