% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sb02mt_nl.R
\name{sb02mt_nl}
\alias{sb02mt_nl}
\title{sb02mt_nl}
\usage{
sb02mt_nl(uplo, n, m, a, b, q, r, l, ldwork)
}
\description{
Conversion of optimal problems with coupling weighting terms to standard problems
}
\examples{
  To compute the following matrices

             -1
      G = B*R  *B',

      -          -1
      A = A - B*R  *L',

      -          -1
      Q = Q - L*R  *L',

  where A, B, Q, R, L, and G are N-by-N, N-by-M, N-by-N, M-by-M,
  N-by-M, and N-by-N matrices, respectively, with Q, R and G
  symmetric matrices.

  When R is well-conditioned with respect to inversion, standard
  algorithms for solving linear-quadratic optimization problems will
  then also solve optimization problems with coupling weighting
  matrix L. Moreover, a gain in efficiency is possible using matrix
  G in the deflating subspace algorithms (see SLICOT Library routine
  SB02OD) or in the Newton's algorithms (see SLICOT Library routine
  SG02CD).

#'
}
\references{
\url{http://slicot.org/objects/software/shared/doc/SB02MT.html}
}
