% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sb02od_n.R
\name{sb02od_n}
\alias{sb02od_n}
\title{sb02od_n}
\usage{
sb02od_n(dico, uplo, jobl, sort, n, m, a, b, q, r, l, tol, ldwork)
}
\description{
Solution of continuous- or discrete-time algebraic Riccati equations (generalized Schur vectors method)
}
\examples{
  To solve for X either the continuous-time algebraic Riccati
  equation
                           -1
     Q + A'X + XA - (L+XB)R  (L+XB)' = 0                       (1)

  or the discrete-time algebraic Riccati equation
                                  -1
     X = A'XA - (L+A'XB)(R + B'XB)  (L+A'XB)' + Q              (2)

  where A, B, Q, R, and L are N-by-N, N-by-M, N-by-N, M-by-M and
  N-by-M matrices, respectively, such that Q = C'C, R = D'D and
  L = C'D; X is an N-by-N symmetric matrix.
  The routine also returns the computed values of the closed-loop
  spectrum of the system, i.e., the stable eigenvalues lambda(1),
  ..., lambda(N) of the corresponding Hamiltonian or symplectic
  pencil, in the continuous-time case or discrete-time case,
  respectively.
                           -1
  Optionally, matrix G = BR  B' may be given instead of B and R.
  Other options include the case with Q and/or R given in a
  factored form, Q = C'C, R = D'D, and with L a zero matrix.

  The routine uses the method of deflating subspaces, based on
  reordering the eigenvalues in a generalized Schur matrix pair.
  A standard eigenproblem is solved in the continuous-time case
  if G is given.

#'
}
\references{
\url{http://slicot.org/objects/software/shared/doc/SB02OD.html}
}
