% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sb03md.R
\name{sb03md}
\alias{sb03md}
\title{sb03md}
\usage{
sb03md(dico, job, fact, trana, n, a, u, c, ldwork)
}
\description{
Solution of continuous- or discrete-time Lyapunov equations and separation estimation
}
\examples{
  To solve for X either the real continuous-time Lyapunov equation

     op(A)'*X + X*op(A) = scale*C                             (1)

  or the real discrete-time Lyapunov equation

     op(A)'*X*op(A) - X = scale*C                             (2)

  and/or estimate an associated condition number, called separation,
  where op(A) = A or A' (A**T) and C is symmetric (C = C').
  (A' denotes the transpose of the matrix A.) A is N-by-N, the right
  hand side C and the solution X are N-by-N, and scale is an output
  scale factor, set less than or equal to 1 to avoid overflow in X.

#'
}
\references{
\url{http://slicot.org/objects/software/shared/doc/SB03MD.html}
}
