% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sb04md.R
\name{sb04md}
\alias{sb04md}
\title{sb04md}
\usage{
sb04md(n, m, a, b, c, ldwork)
}
\description{
Solution of continuous-time Sylvester equations (Hessenberg-Schur method)
}
\examples{
  To solve for X the continuous-time Sylvester equation

     AX + XB = C

  where A, B, C and X are general N-by-N, M-by-M, N-by-M and
  N-by-M matrices respectively.

#'
}
\references{
\url{http://slicot.org/objects/software/shared/doc/SB04MD.html}
}
