% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sb10ad.R
\name{sb10ad}
\alias{sb10ad}
\title{sb10ad}
\usage{
sb10ad(
  job,
  n,
  m,
  np,
  ncon,
  nmeas,
  gamma,
  a,
  b,
  c,
  d,
  gtol,
  actol,
  liwork,
  ldwork
)
}
\description{
H-infinity optimal controller using modified Glover's and Doyle's formulas (continuous-time)
}
\examples{
  To compute the matrices of an H-infinity optimal n-state
  controller

           | AK | BK |
       K = |----|----|,
           | CK | DK |

  using modified Glover's and Doyle's 1988 formulas, for the system

           | A  | B1  B2  |   | A | B |
       P = |----|---------| = |---|---|
           | C1 | D11 D12 |   | C | D |
           | C2 | D21 D22 |

  and for the estimated minimal possible value of gamma with respect
  to GTOL, where B2 has as column size the number of control inputs
  (NCON) and C2 has as row size the number of measurements (NMEAS)
  being provided to the controller, and then to compute the matrices
  of the closed-loop system

           | AC | BC |
       G = |----|----|,
           | CC | DC |

  if the stabilizing controller exists.

  It is assumed that

  (A1) (A,B2) is stabilizable and (C2,A) is detectable,

  (A2) D12 is full column rank and D21 is full row rank,

  (A3) | A-j*omega*I  B2  | has full column rank for all omega,
       |    C1        D12 |

  (A4) | A-j*omega*I  B1  |  has full row rank for all omega.
       |    C2        D21 |

#'
}
\references{
\url{http://slicot.org/objects/software/shared/doc/SB10AD.html}
}
