% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sb10dd.R
\name{sb10dd}
\alias{sb10dd}
\title{sb10dd}
\usage{
sb10dd(n, m, np, ncon, nmeas, gamma, a, b, c, d, tol, ldwork)
}
\description{
H-infinity (sub)optimal controller for a discrete-time system
}
\examples{
  To compute the matrices of an H-infinity (sub)optimal n-state
  controller

                        | AK | BK |
                    K = |----|----|,
                        | CK | DK |

  for the discrete-time system

                | A  | B1  B2  |   | A | B |
            P = |----|---------| = |---|---|
                | C1 | D11 D12 |   | C | D |
                | C2 | D21 D22 |

  and for a given value of gamma, where B2 has as column size the
  number of control inputs (NCON) and C2 has as row size the number
  of measurements (NMEAS) being provided to the controller.

  It is assumed that

  (A1) (A,B2) is stabilizable and (C2,A) is detectable,

  (A2) D12 is full column rank and D21 is full row rank,

            j*Theta
  (A3) | A-e       *I  B2  | has full column rank for all
       |    C1         D12 |

       0 <= Theta < 2*Pi ,

            j*Theta
  (A4) | A-e       *I  B1  |  has full row rank for all
       |    C2         D21 |

       0 <= Theta < 2*Pi .

#'
}
\references{
\url{http://slicot.org/objects/software/shared/doc/SB10DD.html}
}
