% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sg03ad.R
\name{sg03ad}
\alias{sg03ad}
\title{sg03ad}
\usage{
sg03ad(dico, job, fact, trans, uplo, n, a, e, q, z, x, ldwork)
}
\description{
Solution of continuous- or discrete-time generalized Lyapunov equations and separation estimation
}
\examples{
  To solve for X either the generalized continuous-time Lyapunov
  equation

          T                T
     op(A)  X op(E) + op(E)  X op(A) = SCALE * Y,                (1)

  or the generalized discrete-time Lyapunov equation

          T                T
     op(A)  X op(A) - op(E)  X op(E) = SCALE * Y,                (2)

  where op(M) is either M or M**T for M = A, E and the right hand
  side Y is symmetric. A, E, Y, and the solution X are N-by-N
  matrices. SCALE is an output scale factor, set to avoid overflow
  in X.

  Estimates of the separation and the relative forward error norm
  are provided.

#'
}
\references{
\url{http://slicot.org/objects/software/shared/doc/SG03AD.html}
}
