% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb01id.R
\name{tb01id}
\alias{tb01id}
\title{tb01id}
\usage{
tb01id(job, n, m, p, maxred, a, b, c)
}
\description{
Balancing a system matrix corresponding to a triplet (A,B,C)
}
\examples{
  To reduce the 1-norm of a system matrix

          S =  ( A  B )
               ( C  0 )

  corresponding to the triple (A,B,C), by balancing. This involves
  a diagonal similarity transformation inv(D)*A*D applied
  iteratively to A to make the rows and columns of
                        -1
               diag(D,I)  * S * diag(D,I)

  as close in norm as possible.

  The balancing can be performed optionally on the following
  particular system matrices

           S = A,    S = ( A  B )    or    S = ( A )
                                               ( C )

#'
}
\references{
\url{http://slicot.org/objects/software/shared/doc/TB01ID.html}
}
