% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb01pd.R
\name{tb01pd}
\alias{tb01pd}
\title{tb01pd}
\usage{
tb01pd(job, equil, n, m, p, a, b, c, tol, ldwork)
}
\description{
Minimal, controllable or observable block Hessenberg realization for a given state-space representation
}
\examples{
  To find a reduced (controllable, observable, or minimal) state-
  space representation (Ar,Br,Cr) for any original state-space
  representation (A,B,C). The matrix Ar is in upper block
  Hessenberg form.

#'
}
\references{
\url{http://slicot.org/objects/software/shared/doc/TB01PD.html}
}
