% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb03ad_l.R
\name{tb03ad_l}
\alias{tb03ad_l}
\title{tb03ad_l}
\usage{
tb03ad_l(equil, n, m, p, a, b, c, d, tol, ldwork)
}
\description{
Left/right polynomial matrix representation of a given state-space representation
}
\examples{
  To find a relatively prime left polynomial matrix representation
  inv(P(s))*Q(s) or right polynomial matrix representation
  Q(s)*inv(P(s)) with the same transfer matrix T(s) as that of a
  given state-space representation, i.e.

     inv(P(s))*Q(s) = Q(s)*inv(P(s)) = T(s) = C*inv(s*I-A)*B + D.

#'
}
\references{
\url{http://slicot.org/objects/software/shared/doc/TB03AD.html}
}
