% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb04ad_c.R
\name{tb04ad_c}
\alias{tb04ad_c}
\title{tb04ad_c}
\usage{
tb04ad_c(n, m, p, a, b, c, d, tol1, tol2, ldwork)
}
\description{
Transfer matrix of a given state-space representation (A,B,C,D)
}
\examples{
  To find the transfer matrix T(s) of a given state-space
  representation (A,B,C,D). T(s) is expressed as either row or
  column polynomial vectors over monic least common denominator
  polynomials.

#'
}
\references{
\url{http://slicot.org/objects/software/shared/doc/TB04AD.html}
}
