% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb05ad_ag.R
\name{tb05ad_ag}
\alias{tb05ad_ag}
\title{tb05ad_ag}
\usage{
tb05ad_ag(n, m, p, a, b, c, ldwork, lzwork)
}
\description{
Frequency response matrix of a given state-space representation (A,B,C)
}
\examples{
  To find the complex frequency response matrix (transfer matrix)
  G(freq) of the state-space representation (A,B,C) given by
                                -1
     G(freq) = C * ((freq*I - A)  ) * B

  where A, B and C are real N-by-N, N-by-M and P-by-N matrices
  respectively and freq is a complex scalar.

#'
}
\references{
\url{http://slicot.org/objects/software/shared/doc/TB05AD.html}
}
