% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc01od_r.R
\name{tc01od_r}
\alias{tc01od_r}
\title{tc01od_r}
\usage{
tc01od_r(m, p, indlim, pcoeff, qcoeff)
}
\description{
Dual of a left/right polynomial matrix representation
}
\examples{
  To find the dual right (left) polynomial matrix representation of
  a given left (right) polynomial matrix representation, where the
  right and left polynomial matrix representations are of the form
  Q(s)*inv(P(s)) and inv(P(s))*Q(s) respectively.

#'
}
\references{
\url{http://slicot.org/objects/software/shared/doc/TC01OD.html}
}
