% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc04ad_r.R
\name{tc04ad_r}
\alias{tc04ad_r}
\title{tc04ad_r}
\usage{
tc04ad_r(m, p, index_bn, pcoeff, qcoeff, n, ldwork)
}
\description{
State-space representation for a given left/right polynomial matrix representation
}
\examples{
  To find a state-space representation (A,B,C,D) with the same
  transfer matrix T(s) as that of a given left or right polynomial
  matrix representation, i.e.

     C*inv(sI-A)*B + D = T(s) = inv(P(s))*Q(s) = Q(s)*inv(P(s)).

#'
}
\references{
\url{http://slicot.org/objects/software/shared/doc/TC04AD.html}
}
