% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/td04ad_r.R
\name{td04ad_r}
\alias{td04ad_r}
\title{td04ad_r}
\usage{
td04ad_r(m, p, index_bn, dcoeff, ucoeff, nr, tol, ldwork)
}
\description{
Minimal state-space representation for a proper transfer matrix
}
\examples{
  To find a minimal state-space representation (A,B,C,D) for a
  proper transfer matrix T(s) given as either row or column
  polynomial vectors over denominator polynomials, possibly with
  uncancelled common terms.

#'
}
\references{
\url{http://slicot.org/objects/software/shared/doc/TD04AD.html}
}
