% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf01md.R
\name{tf01md}
\alias{tf01md}
\title{tf01md}
\usage{
tf01md(n, m, p, ny, a, b, c, d, u, x)
}
\description{
Output response sequence of a linear time-invariant discrete-time system
}
\examples{
  To compute the output sequence of a linear time-invariant
  open-loop system given by its discrete-time state-space model
  (A,B,C,D), where A is an N-by-N general matrix.

  The initial state vector x(1) must be supplied by the user.

#'
}
\references{
\url{http://slicot.org/objects/software/shared/doc/TF01MD.html}
}
