% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf01rd.R
\name{tf01rd}
\alias{tf01rd}
\title{tf01rd}
\usage{
tf01rd(na, nb, nc, n, a, b, c, ldwork)
}
\description{
Markov parameters of a multivariable system from the state-space representation
}
\examples{
  To compute N Markov parameters M(1), M(2),..., M(N) from the
  parameters (A,B,C) of a linear time-invariant system, where each
  M(k) is an NC-by-NB matrix and k = 1,2,...,N.

  All matrices are treated as dense, and hence TF01RD is not
  intended for large sparse problems.

#'
}
\references{
\url{http://slicot.org/objects/software/shared/doc/TF01RD.html}
}
