% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tg01ad.R
\name{tg01ad}
\alias{tg01ad}
\title{tg01ad}
\usage{
tg01ad(job, l, n, m, p, thresh, a, e, b, c)
}
\description{
Balancing the matrices of the system pencil corresponding to a descriptor triple (A-lambda E,B,C)
}
\examples{
  To balance the matrices of the system pencil

          S =  ( A  B ) - lambda ( E  0 ) :=  Q - lambda Z,
               ( C  0 )          ( 0  0 )

  corresponding to the descriptor triple (A-lambda E,B,C),
  by balancing. This involves diagonal similarity transformations
  (Dl*A*Dr - lambda Dl*E*Dr, Dl*B, C*Dr) applied to the system
  (A-lambda E,B,C) to make the rows and columns of system pencil
  matrices

               diag(Dl,I) * S * diag(Dr,I)

  as close in norm as possible. Balancing may reduce the 1-norms
  of the matrices of the system pencil S.

  The balancing can be performed optionally on the following
  particular system pencils

           S = A-lambda E,

           S = ( A-lambda E  B ),    or

           S = ( A-lambda E ).
               (     C      )

#'
}
\references{
\url{http://slicot.org/objects/software/shared/doc/TG01AD.html}
}
