% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tg01fd_uu.R
\name{tg01fd_uu}
\alias{tg01fd_uu}
\title{tg01fd_uu}
\usage{
tg01fd_uu(joba, l, n, m, p, a, e, b, c, q, z, tol, ldwork)
}
\description{
Orthogonal reduction of a descriptor system to a SVD-like coordinate form
}
\examples{
  To compute for the descriptor system (A-lambda E,B,C)
  the orthogonal transformation matrices Q and Z such that the
  transformed system (Q'*A*Z-lambda Q'*E*Z, Q'*B, C*Z) is
  in a SVD-like coordinate form with

               ( A11  A12 )             ( Er  0 )
      Q'*A*Z = (          ) ,  Q'*E*Z = (       ) ,
               ( A21  A22 )             (  0  0 )

  where Er is an upper triangular invertible matrix.
  Optionally, the A22 matrix can be further reduced to the form

               ( Ar  X )
         A22 = (       ) ,
               (  0  0 )

  with Ar an upper triangular invertible matrix, and X either a full
  or a zero matrix.
  The left and/or right orthogonal transformations performed
  to reduce E and A22 can be optionally accumulated.

#'
}
\references{
\url{http://slicot.org/objects/software/shared/doc/TG01FD.html}
}
